fix: r1-fix — distribute tests inline per sub-module, delete tests.rs
37 tests distributed to their locality sub-modules:
- tree.rs: 5 tests (ratio_bounds, pane_count, swap)
- split.rs: 10 tests (ratio_to_size, split_child_areas, hit-testing)
- pane.rs: 2 tests (composing state)
- tab.rs: 14 tests (title_bar, split_at, close, cycle, layout, detach)
- mod.rs: 6 tests (Layout next_tab, move_pane, find_agent)

All sub-files ≤492 LOC (well under 700 closure metric).
